move helper funcs to their respective type's files
