+++ src/HPolyhedron.jl
@@ -392,117 +392,6 @@ function remove_redundant_constraints!(P::HPoly{N};
     remove_redundant_constraints!(P.constraints, backend=backend)
 end
 
-"""
-    linear_map(M::AbstractMatrix{N},
-               P::PT;
-               check_invertibility::Bool=true,
-               cond_tol::Number=DEFAULT_COND_TOL,
-               use_inv::Bool=!issparse(M)
-               ) where {N<:Real, PT<:AbstractPolyhedron{N}}
-
-Concrete linear map of a polyhedron in constraint representation.
-
-### Input
-
-- `M` -- matrix
-- `P` -- polyhedron in constraint representation
-- `check_invertibility` -- (optional, deault: `true`) check if the linear map is
-                           invertible, in which case this function uses the matrix
-                           inverse; if the invertibility check fails, or if
-                           this flag is set to `false`, use the vertex representation
-                           to compute the linear map (see below for details)
-- `cond_tol` -- (optional) tolerance of matrix condition (used to check whether
-                the matrix is invertible)
-- `use_inv`  -- (optional, default: `false` if `M` is sparse and `true`
-                otherwise) whether to compute the full left division through
-                `inv(M)`, or to use the left division for each vector; see below 
-
-### Output
-
-A polyhedron of the same type as the input if `M` is invertible; otherwise,
-a polytope in vertex representation.
-
-### Algorithm
-
-This function implements essentially two methods for the linear map.
-If the matrix ``M`` is invertible (which we check with a sufficient condition),
-then ``y = M x`` implies ``x = \\text{inv}(M) y`` and we transform the
-constraint system ``A x ≤ b`` to ``A \\text{inv}(M) y ≤ b``. The other method
-consists of passing to the vertex representation of the polyhedron and applying the
-linear map to each vertex.
-
-To switch off the check for invertibility, set the option
-`check_invertibility=false`. If `M` is not invertible and the polyhedron
-is unbounded, this function returns an exception.
-
-The option `use_inv` lets the user control - in case `M` is invertible - if
-the full matrix inverse is computed, or only the left division on the normal
-vectors. Note that this helps as a workaround when `M` is a sparse matrix, since
-the `inv` function is not available for sparse matrices. In this case, either
-use the option `use_inv=false` or convert the type of `M` as in `linear_map(Matrix(M), P)`.
-"""
-function linear_map(M::AbstractMatrix{N},
-                    P::AbstractPolyhedron{N};
-                    check_invertibility::Bool=true,
-                    cond_tol::Number=DEFAULT_COND_TOL,
-                    use_inv::Bool=!issparse(M)
-                   ) where {N<:Real}
-    @assert dim(P) == size(M, 2) "a linear map of size $(size(M)) cannot be applied to a set of dimension $(dim(P))"
-
-    if !check_invertibility || !isinvertible(M; cond_tol=cond_tol)
-        return _linear_map_vrep(M, P)
-    end
-
-    # matrix M is invertible => the normal vectors are vec(c.a' * inv(M))
-    return _linear_map_hrep(M, P, use_inv)
-end
-
-function _linear_map_vrep(M::AbstractMatrix{N}, P::AbstractPolyhedron{N}) where {N<:Real}
-    if !isbounded(P)
-        throw(ArgumentError("the linear map in vertex representation for an unbounded set is not defined"))
-    else
-        # since P is bounded, we pass an HPolytope and then convert it to vertex representation
-        P = tovrep(HPolytope(constraints_list(P), check_boundedness=false))
-    end
-    return _linear_map_vrep(M, P)
-end
-
-function _linear_map_hrep(M::AbstractMatrix{N}, P::AbstractPolyhedron{N},
-                          use_inv::Bool) where {N<:Real}
-    constraints = _linear_map_hrep_helper(M, P, use_inv)
-    return HPolyhedron(constraints)
-end
-
-function _linear_map_hrep(M::AbstractMatrix{N}, P::AbstractPolytope{N},
-                          use_inv::Bool) where {N<:Real}
-    constraints = _linear_map_hrep_helper(M, P, use_inv)
-    return dim(P) == 2 ? HPolygon(constraints) : HPolytope(constraints)
-end
-
-function _linear_map_hrep(M::AbstractMatrix{N}, P::HalfSpace{N},
-                          use_inv::Bool) where {N<:Real}
-    constraint = _linear_map_hrep_helper(M, P, use_inv)
-    return HalfSpace(constraint[1].a, constraint[1].b)
-end
-
-function _linear_map_hrep_helper(M::AbstractMatrix{N}, P::AbstractPolyhedron{N},
-                                 use_inv::Bool) where {N<:Real}
-    constraints_P = constraints_list(P)
-    constraints_MP = similar(constraints_P)
-    if use_inv
-        invM = inv(M)
-        @inbounds for (i, c) in enumerate(constraints_P)
-            constraints_MP[i] = LinearConstraint(vec(c.a' * invM), c.b)
-        end
-    else
-        # take left division for each constraint c, transpose(M) \ c.a
-        @inbounds for (i, c) in enumerate(constraints_P)
-            constraints_MP[i] = LinearConstraint(_At_ldiv_B(M, c.a), c.b)
-        end
-    end
-    return constraints_MP
-end
-
 # ========================================================
 # External methods that require Polyhedra.jl to be loaded
 # ========================================================
