+++ src/HalfSpace.jl
@@ -416,3 +416,8 @@ function is_tighter_same_dir_2D(c1::LinearConstraint{N},
     end
     return lt(c1.b, c1.a[1] / c2.a[1] * c2.b)
 end
+
+function _linear_map_hrep(M::AbstractMatrix{N}, P::HalfSpace{N}, use_inv::Bool) where {N<:Real}
+    constraint = _linear_map_hrep_helper(M, P, use_inv)
+    return HalfSpace(constraint[1].a, constraint[1].b)
+end
